Capital Budgeting - A complete case Study
In this course challenge we had to perform a capital budgeting exercise from scratch. The financial model you built in this challenge was used to understand whether an investment in a certain power plant was feasible or not from a financial point of view. Calculated Project's Feasibility in terms of NPV, IRR and performed a Sensitivity Analysis.